feat(slider): support custom labels
