+++ packages/elemental-theme/src/custom-elements/ef-slider-marker.less
@@ -1,0 +1,21 @@
+:host {
+  width: @slider-step-width;
+  height: (@slider-step-width * 2);
+  top: (@slider-step-width * 2);
+  user-select: none;
+
+  [part='marker'] {
+    background-color: var(--marker-color, @slider-thumb-color);
+  }
+
+  [part='label'] {
+    top: 15px;
+    transform: translateX(calc(-50% + @slider-step-width / 2));
+  }
+  &[label-align='right'] [part='label'] {
+    transform: translateX(calc(-100% + @slider-step-width));
+  }
+  &[label-align='left'] [part='label'] {
+    transform: none;
+  }
+}
