+++ packages/elements/src/slider/__test__/slider.event.test.js
@@ -1,7 +1,7 @@
 import '@refinitiv-ui/elements/slider';
 
 import '@refinitiv-ui/elemental-theme/light/ef-slider';
-import { elementUpdated, expect, fixture, oneEvent } from '@refinitiv-ui/test-helpers';
+import { elementUpdated, expect, fixture, isNear, oneEvent } from '@refinitiv-ui/test-helpers';
 
 import { calculateValue, tabSliderPosition } from './utils.js';
 
@@ -66,8 +66,8 @@ describe('slider/Events', function () {
     await oneEvent(slider, 'mousedown');
     expect(isDragging(el)).to.be.true;
     expect(el.from).to.equal(calculateValue(el, 150).toFixed(0));
-
-    expect(el.to).to.equal(calculateValue(el, window.innerWidth - 100).toString());
+    // Android: Need to test by comparing different value
+    expect(isNear(el.to, calculateValue(el, window.innerWidth - 100).toString(), 1, true)).to.be.true;
 
     setTimeout(() =>
       window.dispatchEvent(
@@ -90,7 +90,8 @@ describe('slider/Events', function () {
     await oneEvent(window, 'mouseup');
     expect(isDragging(el)).to.be.false;
     expect(el.from).to.equal(calculateValue(el, 150).toFixed(0));
-    expect(el.to).to.equal(calculateValue(el, window.innerWidth - 90).toString());
+    // Android: Need to test by comparing different value
+    expect(isNear(el.to, calculateValue(el, window.innerWidth - 90).toString(), 1, true)).to.be.true;
   });
 
   it('Drag "from" thumb slider to end of right.', async function () {
@@ -143,7 +144,9 @@ describe('slider/Events', function () {
     await oneEvent(slider, 'mousedown');
     expect(isDragging(el)).to.be.true;
     expect(el.from).to.equal('0');
-    expect(el.to).to.equal(calculateValue(el, window.innerWidth - 100).toString());
+
+    // Android: Need to test by comparing different value
+    expect(isNear(el.to, calculateValue(el, window.innerWidth - 100).toString(), 1, true)).to.be.true;
 
     setTimeout(() => window.dispatchEvent(new MouseEvent('mousemove', { clientX: 0, clientY: 0 })));
     await oneEvent(window, 'mousemove');
@@ -1127,7 +1130,7 @@ describe('slider/Events', function () {
   it('Event value-changed should fires when value property was set via api and drag the slider back to previous value', async function () {
     expect(el.value).to.equal('0');
     el.value = 10;
-    await elementUpdated();
+    await elementUpdated(el);
     expect(el.value).to.equal('10');
 
     let callCountValue = 0;
@@ -1186,7 +1189,7 @@ describe('slider/Events', function () {
     expect(el.to).to.equal('100');
 
     el.from = 10;
-    await elementUpdated();
+    await elementUpdated(el);
     expect(el.from).to.equal('10');
 
     let callCountValue = 0;
@@ -1236,7 +1239,7 @@ describe('slider/Events', function () {
     expect(el.to).to.equal('100');
 
     el.to = 80;
-    await elementUpdated();
+    await elementUpdated(el);
     expect(el.to).to.equal('80');
 
     let callCountValue = 0;
@@ -1273,6 +1276,7 @@ describe('slider/Events', function () {
     // Check call fire event
     expect(callCountValue).to.equal(1);
   });
+
   it('Should fires input event when dragging from thumb slider with mouse', async function () {
     // Drag 'value' from 0 to 10
     const dragPositionStart = tabSliderPosition(el, 0);
