+++ packages/halo-theme/src/custom-elements/ef-slider-marker.less
@@ -1,0 +1,7 @@
+@import '@refinitiv-ui/elemental-theme/src/custom-elements/ef-slider-marker';
+
+:host {
+  [part='marker'] {
+    background-color: var(--marker-color, @slider-track-fill-color);
+  }
+}
