+++ packages/halo-theme/src/custom-elements/ef-slider.less
@@ -39,7 +39,7 @@
   }
 
   &:not([disabled]) [part='track-fill'] {
-    background-color: var(--thumb-color, @slider-track-fill-color);
+    background-color: var(--progress-color, var(--thumb-color, @slider-track-fill-color));
   }
 
   &:not([disabled]):not([readonly]):not(.grabbable) [part='thumb']:hover {
